レイヤーごとの受け渡し構造体の実装

+++ pkg/cago/model/user.go
@@ -7,7 +7,7 @@ import (
 
 // User ユーザーモデル
 type User struct {
-	ID        int16
+	ID        int32
 	Name      string
 	Token     string
 	CreatedAt time.Time
